start refactoring contexts
