+++ src/utils/jquery.js
@@ -210,13 +210,14 @@ $.fn.notIn = function(element) {
 /**
  * Capture the pinch gesture on trackpads, which is the same as scrolling
  * with the ctrl key pressed (http://stackoverflow.com/a/28685082/4966649).
+ * Can be unbound by calling $(element).off(".pinch").
  *
  * @param {function(Event)} callback - The callback to run when a pinch is
  *   detected. The event will have deltaY defined, which represents the velocity
  *   of the pinch. A positive deltaY means the pinch is getting wider apart.
  */
 $.fn.pinch = function(callback) {
-    return this.on("wheel", function(e) {
+    return this.on("wheel.pinch", function(e) {
         if (e.ctrlKey) {
             e.deltaY = -e.originalEvent.deltaY;
             callback(e);
